backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sky-bound-journeys-backend"
version = "0.1.0"
description = "Backend for SkyBound Journeys flight booking application"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
authors = [
    {name = "SkyBound Journeys Team", email = "dev@example.com"}
]
dependencies = [
    "fastapi[all]>=0.95.0",
    "fastapi-mail>=1.5.0",
    "httpx>=0.24.0",
    "passlib>=1.7.4",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.22.0",
    "sse-starlette>=1.6.0",
    "supabase>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.10.0",
    "asyncpg>=0.27.0",
    "greenlet>=2.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.2.0",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools]
packages = ["app", "alembic"]
